fix: derive purpose from xpriv and adjust parity in sk on bridge-client init
